Make verifier analyzer contract explicit
