+++ tests/promote-final-workflow.test.ts
@@ -21,6 +21,7 @@ describe("final promotion workflow", () => {
 
   it("attests candidate artifacts with a credential-free checkout", async () => {
     const workflow = await readFile(releaseWorkflowPath, "utf8");
+    const githubRefName = shellVariable("GITHUB_REF_NAME");
     expect(workflow).toContain("permissions:\n  contents: read");
     expect(workflow).toContain("build:");
     expect(workflow).toContain("release:\n    needs: build");
@@ -29,7 +30,7 @@ describe("final promotion workflow", () => {
     expect(workflow).toContain("attestations: write");
     expect(workflow).toContain("actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02");
     expect(workflow).toContain("actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093");
-    expect(workflow).toContain('git rev-parse "refs/tags/${GITHUB_REF_NAME}^{commit}"');
+    expect(workflow).toContain(`git rev-parse "refs/tags/${githubRefName}^{commit}"`);
     expect(workflow).toContain('= "$GITHUB_SHA"');
     expect(workflow).toContain("actions/attest@f7c74d28b9d84cb8768d0b8ca14a4bac6ef463e6");
     expect(workflow).toContain("subject-path:");
